Fixed a VS2022 resource compiler issue & resolved some performance issues by better caching repeated metadata calls

+++ in_sidplay2/ThreadSidDecoder.cpp
@@ -585,7 +585,7 @@ void CThreadSidDecoder::DoSeek()
 	}
 	else
 	{
-		timesek = (m_seekNeedMs - m_playTimems) / 1000;
+		timesek = (int)(m_seekNeedMs - m_playTimems) / 1000;
 		if (timesek <= 0) return;
 	}
